Use a buffer on stack to store addresses
Before this change, the performance of the system call was hugely dependent on the size of the buffer provided by the application from userspace regardless of the number of modified addresses there actually were. Oftentimes in real-world usage, the number of addresses returned is small. Using a small buffer and repeatedly copying it out to userspace is far more performant in common cases (i.e., not too many addresses being returned) than malloc'ing a huge buffer and copying it out.